Do not require autorun
Requiring autorun makes Minitest to run the tests in an `at_exit` block, which
may or may not be what the user wants. This change makes the gem less
opinionated and gives the developer using the gem decide when and how to run the
tests.

Related https://github.com/kern/minitest-reporters/pull/100