Show location of test failures
While exception messages show a the file name/line number where the
exception was thrown, actual test failures (e.g. `refute true`) were
not displaying the location of the exception.

To expose the problem, I ran the bad test gallery with stock minitest
pride output. The __FILE__ and __LINE__ where the failed assert occured
were showing up. However, when I switched to
`Minitest::Reporters::DefaultReporter`, the failure location was gone.

I also added the ability to easily run the galleries with minitest pride
output as a way of testing the formatting of `DefaultReporter` against
stock minitest in the future.