Always output run options for stdout reporters
This is important because of Minitest's random ordering-- if there are
bugs that occur in some orderings, being able to run with the same seed
to reproduce the same ordering is essential.

This was added to minitest-reporters once before, in df651b0, but looks
like it didn't survive the 1.0 refactoring :) Discussion when that was
added: https://github.com/kern/minitest-reporters/pull/55/files

Did not add to the junit reporter since its output mostly goes to XML
files, not stdout, and if you use it in conjunction with the other
reporters that typically print to stdout, you will see this twice. If
you use multiple stdout reporters, you're going to get a lot of
duplicate info as well.